create context and custon hook to initialStateUser
