+++ src/routes/App.jsx
@@ -7,19 +7,23 @@ import Home from '@/views/Home'
 import Search from '@/views/Search'
 import NotFound from '@/views/Errors/NotFound.jsx'
 
+import AppContext from '@/context/AppContext'
+import useInitialStateUsers from '@/hooks/useInitialStateUsers'
+
 const App = () => {
+  const initalState = useInitialStateUsers()
   return (
-  <>
-    <BrowserRouter>
-      <Layout>
-        <Routes>
-          <Route exact path="/" element={<Home />} />
-          <Route exact path="/user" element={<Search />} />
-          <Route path="*" element={<NotFound />} />
-        </Routes>
-      </Layout>
-    </BrowserRouter>
-  </>
+    <AppContext.Provider value={initalState}>
+      <BrowserRouter>
+        <Layout>
+          <Routes>
+            <Route exact path="/" element={<Home />} />
+            <Route exact path="/user" element={<Search />} />
+            <Route path="*" element={<NotFound />} />
+          </Routes>
+        </Layout>
+      </BrowserRouter>
+    </AppContext.Provider>
   )
 }
